moved types to type folder
