--- src/js/engagement_view/src/components/login/checkCookieWhenRouting.tsx
@@ -1,9 +1,0 @@
-import { RouteState } from '../../types/CustomTypes';
-
-
-export const defaultRouteState = (): RouteState => {
-    return {
-        curPage: localStorage.getItem("grapl_curPage") || "login",
-        lastCheckLoginCheck: Date.now()
-    }
-}
