--- src/js/engagement_view/src/css/sidebar.css
@@ -1,4 +1,0 @@
-sidebar{
-    background-color:red;
-    /* width: 33%; */
-}
