+++ src/js/engagement_view/src/index.css
@@ -3,10 +3,6 @@
 body {
     margin: 0;
     font-family: 'Roboto Condensed', sans-serif;
-    /* font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', 'Oxygen',
-    'Ubuntu', 'Cantarell', 'Fira Sans', 'Droid Sans', 'Helvetica Neue',
-    sans-serif; */
-
     -webkit-font-smoothing: antialiased;
     -moz-osx-font-smoothing: grayscale;
     background-color:#303030;
