--- src/js/engagement_view/src/services/getApiURLs.tsx
@@ -1,11 +1,0 @@
-export const getAuthEdge = () => {
-    return "http://"+window.location.hostname+":8900/"
-};
-
-export const getGraphQlEdge = () => {
-    return "http://"+window.location.hostname+":5000/"
-};
-
-export const getModelPluginEdge = () => {
-    return "http://"+window.location.hostname+":8123/"
-};
