+++ src/js/engagement_view/src/services/graphQL/utils_GraphQlEndpointRetrieveGraph/expandScopeQuery.tsx
@@ -1,4 +1,3 @@
-
 export const expandScopeQuery = (lensName: string) => {
     console.log("expanding scope for: ", lensName);
